Fixed memory mapped file deletion
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@26196 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: fc53f0113ab72c8aaf9f11ea1c1679a17b2437f2

+++ common/sensormanagement.common/src/de/uniol/inf/is/odysseus/sensormanagement/common/utilities/MemoryMappedFile.java
@@ -108,8 +108,9 @@ public class MemoryMappedFile
 		if (direction == Direction.READ)
 			stopThread();
 		
-		memoryMappedFile.close();		
-		new File(filename).delete();
+		memoryMappedFile.close();
+		if (deleteFile)
+			new File(filename).delete();
 	}
 
 	public void write(ByteBuffer message) throws IOException
